refactor: refactor securityValidator and Keystore implementations, streamline module structure
