+++ src/security/mod.rs
@@ -1,0 +1,6 @@
+pub mod validator;
+
+#[cfg(test)]
+mod tests;
+
+pub use validator::SecurityValidator;
